+++ core/app/models/spree/promotion/rules/taxon.rb
@@ -7,6 +7,9 @@ module Spree
         has_many :taxons, through: :promotion_rule_taxons, class_name: 'Spree::Taxon'
 
         MATCH_POLICIES = %w(any all)
+
+        validates_inclusion_of :preferred_match_policy, in: MATCH_POLICIES
+
         preference :match_policy, :string, default: MATCH_POLICIES.first
 
         def applicable?(promotable)
@@ -14,13 +17,21 @@ module Spree
         end
 
         def eligible?(order, _options = {})
-          if preferred_match_policy == 'all'
-            unless (taxons.to_a - taxons_in_order_including_parents(order)).empty?
+          order_taxons = taxons_in_order_including_parents(order)
+
+          case preferred_match_policy
+          when 'all'
+            unless (taxons.to_a - order_taxons).empty?
               eligibility_errors.add(:base, eligibility_error_message(:missing_taxon))
             end
-          else
-            order_taxons = taxons_in_order_including_parents(order)
+          when 'any'
             unless taxons.any?{ |taxon| order_taxons.include? taxon }
+              eligibility_errors.add(:base, eligibility_error_message(:no_matching_taxons))
+            end
+          else
+            # Change this to an exception in a future version of Solidus
+            warn_invalid_match_policy(assume: 'any')
+            unless taxons.any? { |taxon| order_taxons.include? taxon }
               eligibility_errors.add(:base, eligibility_error_message(:no_matching_taxons))
             end
           end
@@ -29,7 +40,14 @@ module Spree
         end
 
         def actionable?(line_item)
-          taxon_product_ids.include? line_item.variant.product_id
+          case preferred_match_policy
+          when 'any', 'all'
+            taxon_product_ids.include?(line_item.variant.product_id)
+          else
+            # Change this to an exception in a future version of Solidus
+            warn_invalid_match_policy(assume: 'any')
+            taxon_product_ids.include?(line_item.variant.product_id)
+          end
         end
 
         def taxon_ids_string
@@ -42,6 +60,14 @@ module Spree
         end
 
         private
+
+        def warn_invalid_match_policy(assume:)
+          ActiveSupport::Deprecation.warn(
+            "#{self.class.name} id=#{id} has unexpected match policy #{preferred_match_policy.inspect}. "\
+            "Interpreting it as '#{assume}'." \
+            "In future versions of Solidus this will be an error."
+          )
+        end
 
         # All taxons in an order
         def order_taxons(order)