+++ core/spec/models/spree/promotion/rules/taxon_spec.rb
@@ -96,5 +96,51 @@ describe Spree::Promotion::Rules::Taxon, type: :model do
         it{ expect(rule).to be_eligible(order) }
       end
     end
+
+    context 'with an invalid match policy' do
+      before do
+        order.products.first.taxons << taxon
+        rule.taxons << taxon
+        rule.preferred_match_policy = 'invalid'
+        rule.save!(validate: false)
+      end
+
+      it 'logs a warning and uses "any" policy' do
+        expect(ActiveSupport::Deprecation).to(
+          receive(:warn).
+          with(/has unexpected match policy "invalid"/)
+        )
+
+        expect(
+          rule.eligible?(order)
+        ).to be_truthy
+      end
+    end
+  end
+
+  describe '#actionable?' do
+    let(:line_item) { order.line_items.first! }
+    let(:order) { create :order_with_line_items }
+    let(:taxon) { create :taxon, name: 'first' }
+
+    before do
+      rule.preferred_match_policy = 'invalid'
+      rule.save!(validate: false)
+      line_item.product.taxons << taxon
+      rule.taxons << taxon
+    end
+
+    context 'with an invalid match policy' do
+      it 'logs a warning and uses "any" policy' do
+        expect(ActiveSupport::Deprecation).to(
+          receive(:warn).
+          with(/has unexpected match policy "invalid"/)
+        )
+
+        expect(
+          rule.actionable?(line_item)
+        ).to be_truthy
+      end
+    end
   end
 end